# cmd/web-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "SaaS Example API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/accounts": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Find returns the existing accounts in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "List accounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter string, example: number = 'SB10000001'",
                        "name": "where",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order columns separated by comma, example: created_at desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit, example: 10",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset, example: 20",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Included Archived, example: false",
                        "name": "include-archived",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Included Customer info, example: false",
                        "name": "include-customer",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Included Branch info, example: false",
                        "name": "include-branch",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Included Sale rep info, example: false",
                        "name": "include-sales-rep",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/account.Response"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Create inserts a new account into the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Create new account.",
                "parameters": [
                    {
                        "description": "Account details",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/account.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/account.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Update updates the specified account in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Update account by ID",
                "parameters": [
                    {
                        "description": "Update fields",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/tenant.AccountUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/archive": {
            "patch": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Archive soft-deletes the specified customer from the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Archive account by ID",
                "parameters": [
                    {
                        "description": "Update fields",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/account.ArchiveRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Read returns the specified account from the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Get account by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tenant.AccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/weberror.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "account.ArchiveRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "985f1746-1d9f-459f-a2d9-fc53ece5ae86"
                }
            }
        },
        "account.CreateRequest": {
            "type": "object",
            "required": [
                "customer_id",
                "type"
            ],
            "properties": {
                "branch_id": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string",
                    "example": "985f1746-1d9f-459f-a2d9-fc53ece5ae86"
                },
                "target": {
                    "type": "number"
                },
                "target_info": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "account.Response": {
            "type": "object",
            "properties": {
                "archived_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                },
                "branch": {
                    "type": "string"
                },
                "branch_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                },
                "customer": {
                    "type": "object",
                    "$ref": "#/definitions/customer.Response"
                },
                "id": {
                    "type": "string",
                    "example": "985f1746-1d9f-459f-a2d9-fc53ece5ae86"
                },
                "number": {
                    "type": "string",
                    "example": "Rocket Launch"
                },
                "sales_rep": {
                    "type": "object",
                    "$ref": "#/definitions/user.UserResponse"
                },
                "sales_rep_id": {
                    "type": "string"
                },
                "target": {
                    "type": "number"
                },
                "target_info": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                }
            }
        },
...
        "customer.Response": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "archived_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                },
                "branch": {
                    "type": "string"
                },
                "branch_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "985f1746-1d9f-459f-a2d9-fc53ece5ae86"
                },
                "name": {
                    "type": "string",
                    "example": "Rocket Launch"
                },
                "phone_number": {
                    "type": "string"
                },
                "sales_rep": {
                    "type": "string"
                },
                "sales_rep_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                }
            }
        },
...
        "tenant.AccountResponse": {
            "type": "object",
            "properties": {
                "address1": {
                    "type": "string",
                    "example": "221 Tatitlek Ave"
                },
                "address2": {
                    "type": "string",
                    "example": "Box #1832"
                },
                "archived_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                },
                "billing_user_id": {
                    "type": "string",
                    "example": "d69bdef7-173f-4d29-b52c-3edc60baf6a2"
                },
                "city": {
                    "type": "string",
                    "example": "Valdez"
                },
                "country": {
                    "type": "string",
                    "example": "USA"
                },
                "created_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                },
                "id": {
                    "type": "string",
                    "example": "c4653bf9-5978-48b7-89c5-95704aebb7e2"
                },
                "name": {
                    "type": "string",
                    "example": "Company Name"
                },
                "region": {
                    "type": "string",
                    "example": "AK"
                },
                "signup_user_id": {
                    "type": "string",
                    "example": "d69bdef7-173f-4d29-b52c-3edc60baf6a2"
                },
                "status": {
                    "type": "object",
                    "$ref": "#/definitions/web.EnumResponse"
                },
                "timezone": {
                    "type": "string",
                    "example": "America/Anchorage"
                },
                "updated_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                },
                "zipcode": {
                    "type": "string",
                    "example": "99686"
                }
            }
        },
        "tenant.AccountUpdateRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "address1": {
                    "type": "string",
                    "example": "221 Tatitlek Ave"
                },
                "address2": {
                    "type": "string",
                    "example": "Box #1832"
                },
                "billing_user_id": {
                    "type": "string",
                    "example": "d69bdef7-173f-4d29-b52c-3edc60baf6a2"
                },
                "city": {
                    "type": "string",
                    "example": "Valdez"
                },
                "country": {
                    "type": "string",
                    "example": "USA"
                },
                "id": {
                    "type": "string",
                    "example": "c4653bf9-5978-48b7-89c5-95704aebb7e2"
                },
                "name": {
                    "type": "string",
                    "example": "Company Name"
                },
                "region": {
                    "type": "string",
                    "example": "AK"
                },
                "signup_user_id": {
                    "type": "string",
                    "example": "d69bdef7-173f-4d29-b52c-3edc60baf6a2"
                },
                "timezone": {
                    "type": "string",
                    "example": "America/Anchorage"
                },
                "zipcode": {
                    "type": "string",
                    "example": "99686"
                }
            }
        },
...
        "user.UserResponse": {
            "type": "object",
            "properties": {
                "archived_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                },
                "created_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                },
                "email": {
                    "type": "string",
                    "example": "gabi@geeksinthewoods.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "Gabi"
                },
                "gravatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "d69bdef7-173f-4d29-b52c-3edc60baf6a2"
                },
                "last_name": {
                    "type": "string",
                    "example": "May"
                },
                "name": {
                    "type": "string",
                    "example": "Gabi"
                },
                "timezone": {
                    "type": "string",
                    "example": "America/Anchorage"
                },
                "updated_at": {
                    "type": "object",
                    "$ref": "#/definitions/web.TimeResponse"
                }
            }
        },
...
        "web.EnumResponse": {
            "type": "object",
            "properties": {
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.EnumOption"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Active Etc"
                },
                "value": {
                    "type": "string",
                    "example": "active_etc"
                }
            }
        },
        "web.TimeResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2019-06-25"
                },
                "kitchen": {
                    "type": "string",
                    "example": "3:00AM"
                },
                "local": {
                    "type": "string",
                    "example": "Tue Jun 25 3:00AM"
                },
                "local_date": {
                    "type": "string",
                    "example": "Tue Jun 25"
                },
                "local_time": {
                    "type": "string",
                    "example": "3:00AM"
                },
                "now_rel_time": {
                    "type": "string",
                    "example": "15 hours from now"
                },
                "now_time": {
                    "type": "string",
                    "example": "5 hours ago"
                },
                "rfc1123": {
                    "type": "string",
                    "example": "Tue, 25 Jun 2019 03:00:53 AKDT"
                },
                "time": {
                    "type": "string",
                    "example": "03:00:53"
                },
                "timezone": {
                    "type": "string",
                    "example": "America/Anchorage"
                },
                "value": {
                    "type": "string",
                    "example": "2019-06-25T03:00:53.284-08:00"
                },
                "value_utc": {
                    "type": "string",
                    "example": "2019-06-25T11:00:53.284Z"
                }
            }
        },
        "weberror.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/weberror.FieldError"
                    }
                },
                "stack_trace": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "weberror.FieldError": {
            "type": "object",
            "properties": {
                "display": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "object"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/v1/oauth/token",
            "scopes": {
                "admin": " Grants administrative privileges with role of admin.",
                "user": " Grants basic privileges with role of user."
            }
        }
    }
}